requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
tzdata==2025.2
asgiref
django
tzdata
google-generativeai
Pillow
opencv-python
pandas
